Revert "renamed computer parsers to laptop parsers (for clearity)"
This reverts commit 1bdfb291907888a9b15e1e8765692fe7e16c2a22.
