+++ DataUpdater/DataUpdater.cs
@@ -1,14 +1,18 @@
-﻿using ItemLibrary;
+﻿using WebParser.ComputerParsers;
+using ItemLibrary;
+using DataContent.ReadingCSV.Services;
 using System.Collections.Generic;
+using PathLibrary;
 using WebParser;
+using DataContent;
 using static ItemLibrary.Categories;
 using System.Linq;
+using System;
 using System.Threading.Tasks;
 using System.Net.Http;
 using Newtonsoft.Json;
 using System.Text;
 using WebParser.SmartphoneParsers;
-using WebParser.LaptopParsers;
 
 namespace DataUpdater
 {
@@ -67,9 +71,9 @@ namespace DataUpdater
 
                     List<Task<List<Computer>>> laptopTasks = new List<Task<List<Computer>>>();
 
-                    laptopTasks.Add(Task.Run(() => new SenukaiLaptopParser().ParseShop()));
-                    laptopTasks.Add(Task.Run(() => new AvitelaLaptopParser().ParseShop()));
-                    laptopTasks.Add(Task.Run(() => new PiguLaptoprParser().ParseShop()));
+                    laptopTasks.Add(Task.Run(() => new SenukaiComputerParser().ParseShop()));
+                    laptopTasks.Add(Task.Run(() => new AvitelaComputerParser().ParseShop()));
+                    laptopTasks.Add(Task.Run(() => new PiguComputerParser().ParseShop()));
                     
                     var laptopData = await Task.WhenAll(laptopTasks);
