+++ ShopParser/ComputerParsers/PiguComputerParser.cs
@@ -10,14 +10,14 @@ using System.Net.Http;
 using System.Threading.Tasks;
 using static ItemLibrary.Categories;
 
-namespace WebParser.LaptopParsers
+namespace WebParser.ComputerParsers
 {
-    public class PiguLaptoprParser : IParser<Computer>
+    public class PiguComputerParser : IParser<Computer>
     {
         private readonly string _url = "https://pigu.lt/lt/kompiuteriai/nesiojami-kompiuteriai?page=1";
         private Lazy<ChromeDriver> _driver;
         private HttpClient _client;
-        public PiguLaptoprParser ()
+        public PiguComputerParser ()
         {
             var options = new ChromeOptions();
             options.AddArguments("--headless");
